+++ upcloud/error.go
@@ -9,6 +9,9 @@ import (
 type Error struct {
 	ErrorCode    string `json:"error_code"`
 	ErrorMessage string `json:"error_message"`
+
+	// HTTP Status code
+	Status int `json:"-"`
 }
 
 // UnmarshalJSON is a custom unmarshaller that deals with