+++ upcloud/service/service.go
@@ -135,17 +135,21 @@ func New(client Client) *Service {
 // Parses an error returned from the client into corresponding error type
 func parseJSONServiceError(err error) error {
 	if clientError, ok := err.(*client.Error); ok {
-		var serviceError error
 		switch clientError.Type {
 		case client.ErrorTypeProblem:
-			serviceError = &upcloud.Problem{}
+			p := &upcloud.Problem{}
+			if err := json.Unmarshal(clientError.ResponseBody, p); err != nil {
+				return fmt.Errorf("received malformed client error: %s", string(clientError.ResponseBody))
+			}
+			return p
 		default:
-			serviceError = &upcloud.Error{}
+			ucError := &upcloud.Error{}
+			if err := json.Unmarshal(clientError.ResponseBody, ucError); err != nil {
+				return fmt.Errorf("received malformed client error: %s", string(clientError.ResponseBody))
+			}
+			ucError.Status = clientError.ErrorCode
+			return ucError
 		}
-		if err := json.Unmarshal(clientError.ResponseBody, serviceError); err != nil {
-			return fmt.Errorf("received malformed client error: %s", string(clientError.ResponseBody))
-		}
-		return serviceError
 	}
 	return err
 }