+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+- HTTP status code field to `upcloud.Error` type
+
 ## [5.0.0]
 
 ### Changed